chore: finishes cloud folder metadata
