+++ src/app/cloud/[team-slug]/[project-slug]/(tabs)/(overview)/client_page.tsx
@@ -1,0 +1,18 @@
+'use client'
+
+import * as React from 'react'
+import { useRouteData } from '@cloud/context'
+
+import { InfraOffline } from './infraOffline'
+import { InfraOnline } from './infraOnline'
+
+export const ProjectOverviewPage = () => {
+  const { project } = useRouteData()
+  const { infraStatus } = project
+
+  if (infraStatus === 'done') {
+    return <InfraOnline />
+  }
+
+  return <InfraOffline />
+}
