+++ src/app/cloud/[team-slug]/client_layout.tsx
@@ -1,0 +1,24 @@
+'use client'
+
+import * as React from 'react'
+import { notFound } from 'next/navigation'
+
+import { useRouteData } from '../context'
+
+type ProjectLayoutType = {
+  children: React.ReactNode
+}
+
+export const TeamPageLayout = ({ children }: ProjectLayoutType) => {
+  const { team, reloadTeam } = useRouteData()
+
+  React.useEffect(() => {
+    reloadTeam()
+  }, [reloadTeam])
+
+  if (team === undefined) return null
+
+  if (team === null) notFound()
+
+  return <React.Fragment>{children}</React.Fragment>
+}
